Func: BE - ImageService/UpsertPostAction tweaked

+++ app/Actions/Blog/UpsertPostAction.php
@@ -3,6 +3,7 @@
 namespace App\Actions\Blog;
 
 use App\Models\Post;
+use App\Models\Gallery;
 use App\Models\Baseinfo;
 use App\Services\ImageService;
 use Illuminate\Http\UploadedFile;
@@ -15,20 +16,27 @@ class UpsertPostAction
 {
     public static function execute(PostData $data, ImageService $imageService, ?Collection $files): Post
     {
+        // if hero_image is not null and existing entry have to be replaced
         if ($files->has('hero_image')) {
-            if ($data->id) {
+            if ($data->id && Post::getEntityById($data->id)->hero_image) {
                 $imageService->deleteHeroImages(Post::getEntityById($data->id)->hero_image);
             }
             $imageService->generateNames($files->get('hero_image'));
-            $data->hero_image = $imageService->storeThumbHeroImages([[200, 200], [640, 480]])
-                    ->storeHeroImages()->generateHeroURL()->filenamesDB;
+            $data->hero_image = $imageService->storeThumbHeroImages([[640, 480]])
+                    ->storeHeroImages(60)->generateHeroURL()->filenamesDB;
+        }
+        // If hero_image is null, but the post and entry exist and we don't want to overwrite entry with null
+        // That's the case when updating galleries on the frontend
+        if (!$files->has('hero_image') && $data->id && Post::getEntityById($data->id)->hero_image) {
+            $data->hero_image = Post::getEntityById($data->id)->hero_image;
         }
 
         $post = Post::updateOrCreate(['id' => $data->id], $data->except('tags', 'categories', 'user')->all());
 
         if ($files->has('images'))  {
             $imageService->deleteGallery($post);
-            $imageService->generateGallery($files->toArray(), $post, [[200, 200], [640, 480]]);
+            Gallery::whereBelongsTo($post)->delete();
+            $imageService->generateGallery($files->toArray(), $post, [[640, 480]]);
         }
 
         if ($data->tags) {
